+++ flink/v1.12/flink/src/main/java/org/apache/iceberg/flink/sink/FlinkSink.java
@@ -27,6 +27,7 @@ import java.util.Map;
 import java.util.function.Function;
 import org.apache.flink.api.common.functions.MapFunction;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
+import org.apache.flink.api.common.typeinfo.Types;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.datastream.DataStreamSink;
 import org.apache.flink.streaming.api.datastream.SingleOutputStreamOperator;
@@ -70,12 +71,6 @@ public class FlinkSink {
 
   private static final String ICEBERG_STREAM_WRITER_NAME = IcebergStreamWriter.class.getSimpleName();
   private static final String ICEBERG_FILES_COMMITTER_NAME = IcebergFilesCommitter.class.getSimpleName();
-  private static final String ICEBERG_STREAM_REWRITER_NAME = IcebergStreamRewriter.class.getSimpleName();
-  private static final String ICEBERG_REWRITE_FILES_COMMITTER_NAME = IcebergRewriteFilesCommitter.class.getSimpleName();
-
-
-  public static final String STREAMING_REWRITE_ENABLE = "flink.rewrite.enable";
-  public static final boolean STREAMING_REWRITE_ENABLE_DEFAULT = false;
 
   private FlinkSink() {
   }
@@ -136,8 +131,6 @@ public class FlinkSink {
     private boolean upsert = false;
     private List<String> equalityFieldColumns = null;
     private String uidPrefix = null;
-    private boolean rewrite = false;
-    private Integer rewriteParallelism = null;
 
     private Builder() {
     }
@@ -274,29 +267,6 @@ public class FlinkSink {
       return this;
     }
 
-    /**
-     * Configuring whether to enable the rewrite operator. The rewrite operator will rewrite committed files between
-     * multiple checkpoints.
-     *
-     * @param enable indicate whether it should rewrite committed files.
-     * @return {@link Builder} to connect the iceberg table.
-     */
-    public Builder rewrite(boolean enable) {
-      this.rewrite = enable;
-      return this;
-    }
-
-    /**
-     * Configuring the rewrite parallel number for iceberg stream rewriter.
-     *
-     * @param newRewriteParallelism the number of parallel iceberg stream rewriter.
-     * @return {@link Builder} to connect the iceberg table.
-     */
-    public Builder rewriteParallelism(int newRewriteParallelism) {
-      this.rewriteParallelism = newRewriteParallelism;
-      return this;
-    }
-
     private <T> DataStreamSink<T> chainIcebergOperators() {
       Preconditions.checkArgument(inputCreator != null,
           "Please use forRowData() or forMapperOutputType() to initialize the input DataStream.");
@@ -325,14 +295,10 @@ public class FlinkSink {
 
       // Add single-parallelism committer that commits files
       // after successful checkpoint or end of input
-      SingleOutputStreamOperator<CommitResult> committerStream = appendCommitter(writerStream);
-
-      // Add parallel rewriter and single-parallelism committer to rewrite committed files
-      // when streaming rewrite is enable.
-      SingleOutputStreamOperator<?> rewriterStream = appendRewriter(committerStream);
+      SingleOutputStreamOperator<Void> committerStream = appendCommitter(writerStream);
 
       // Add dummy discard sink
-      return appendDummySink(rewriterStream);
+      return appendDummySink(committerStream);
     }
 
     /**
@@ -361,7 +327,7 @@ public class FlinkSink {
     }
 
     @SuppressWarnings("unchecked")
-    private <T> DataStreamSink<T> appendDummySink(SingleOutputStreamOperator<?> committerStream) {
+    private <T> DataStreamSink<T> appendDummySink(SingleOutputStreamOperator<Void> committerStream) {
       DataStreamSink<T> resultStream = committerStream
           .addSink(new DiscardingSink())
           .name(operatorName(String.format("IcebergSink %s", this.table.name())))
@@ -372,42 +338,10 @@ public class FlinkSink {
       return resultStream;
     }
 
-    private SingleOutputStreamOperator<?> appendRewriter(
-        SingleOutputStreamOperator<CommitResult> committerStream) {
-
-      boolean streamingRewriteEnable = rewrite || PropertyUtil.propertyAsBoolean(table.properties(),
-          STREAMING_REWRITE_ENABLE, STREAMING_REWRITE_ENABLE_DEFAULT);
-      if (!streamingRewriteEnable) {
-        return committerStream;
-      }
-
-      IcebergStreamRewriter fileRewriter = new IcebergStreamRewriter(tableLoader);
-      SingleOutputStreamOperator<RewriteResult> rewrittenStream = committerStream
-          .keyBy(CommitResult::partition)
-          .transform(operatorName(ICEBERG_STREAM_REWRITER_NAME), TypeInformation.of(RewriteResult.class), fileRewriter);
-      if (rewriteParallelism != null) {
-        rewrittenStream = rewrittenStream.setParallelism(rewriteParallelism);
-      }
-      if (uidPrefix != null) {
-        rewrittenStream = rewrittenStream.uid(uidPrefix + "-rewriter");
-      }
-
-      IcebergRewriteFilesCommitter rewriteFilesCommitter = new IcebergRewriteFilesCommitter(tableLoader);
-      SingleOutputStreamOperator<Void> committedStream = rewrittenStream.transform(
-          operatorName(ICEBERG_REWRITE_FILES_COMMITTER_NAME), TypeInformation.of(Void.class), rewriteFilesCommitter)
-          .setParallelism(1)
-          .setMaxParallelism(1);
-      if (uidPrefix != null) {
-        committedStream.uid(uidPrefix + "-rewrite-committer");
-      }
-      return committedStream;
-    }
-
-    private SingleOutputStreamOperator<CommitResult> appendCommitter(
-        SingleOutputStreamOperator<WriteResult> writerStream) {
+    private SingleOutputStreamOperator<Void> appendCommitter(SingleOutputStreamOperator<WriteResult> writerStream) {
       IcebergFilesCommitter filesCommitter = new IcebergFilesCommitter(tableLoader, overwrite);
-      SingleOutputStreamOperator<CommitResult> committerStream = writerStream
-          .transform(operatorName(ICEBERG_FILES_COMMITTER_NAME), TypeInformation.of(CommitResult.class), filesCommitter)
+      SingleOutputStreamOperator<Void> committerStream = writerStream
+          .transform(operatorName(ICEBERG_FILES_COMMITTER_NAME), Types.VOID, filesCommitter)
           .setParallelism(1)
           .setMaxParallelism(1);
       if (uidPrefix != null) {
